Move classes into folder
