+++ classes/GraphVerticalBar.js
